feat: add HttpResolverAlpha
